+++ vistle/module/general/Tracer/Tracer.cpp
@@ -381,7 +381,7 @@ bool Tracer::reduce(int timestep) {
        }
        //setIntParameter("no_startp", numpoints);
        if (rank() == 0 && numpoints != n0*n1 && !m_numStartpointsPrinted) {
-           sendInfo("actually using %d*%d=%d start points", n0, n1, n0*n1);
+           sendInfo("actually using %lu*%lu=%lu start points", (unsigned long)n0, (unsigned long)n1, (unsigned long)n0*n1);
            m_numStartpointsPrinted = true;
        }
        numpoints = n0*n1;
